+++ rider/src/main/kotlin/com/jetbrains/rider/plugins/unity/debugger/breakpoints/UnityPausepointBreakpointType.kt
@@ -1,9 +1,14 @@
 package com.jetbrains.rider.plugins.unity.debugger.breakpoints
 
+import com.intellij.openapi.actionSystem.AnAction
+import com.intellij.openapi.actionSystem.CommonDataKeys
+import com.intellij.openapi.project.DumbAwareAction
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.vfs.VirtualFile
+import com.intellij.xdebugger.XDebugSession
 import com.intellij.xdebugger.breakpoints.SuspendPolicy
 import com.intellij.xdebugger.breakpoints.XLineBreakpoint
+import com.intellij.xdebugger.impl.breakpoints.XBreakpointUtil
 import com.jetbrains.rider.debugger.breakpoint.DotNetLineBreakpointProperties
 import com.jetbrains.rider.debugger.breakpoint.DotNetLineBreakpointType
 import icons.UnityIcons
@@ -40,4 +45,20 @@ class UnityPausepointBreakpointType : DotNetLineBreakpointType(Id, Title) {
     override fun getDefaultSuspendPolicy() = SuspendPolicy.NONE
 
     override fun getVisibleStandardPanels(): EnumSet<StandardPanels> = EnumSet.of(StandardPanels.DEPENDENCY)
+
+    override fun getAdditionalPopupMenuActions(breakpoint: XLineBreakpoint<DotNetLineBreakpointProperties>, currentSession: XDebugSession?): MutableList<out AnAction> {
+        val action = DumbAwareAction.create(UnityPausepointConstants.convertToLineBreakpointText) {
+            val dataContext = it.dataContext
+            val editor = CommonDataKeys.EDITOR.getData(dataContext) ?: return@create
+            val project = editor.project ?: return@create
+
+            // This finds the breakpoint at the current line. We're an alt+enter action, so that's fine
+            val pair = XBreakpointUtil.findSelectedBreakpoint(project, editor)
+            if (breakpoint.properties is DotNetLineBreakpointProperties && pair.second == breakpoint) {
+                @Suppress("UNCHECKED_CAST")
+                convertToLineBreakpoint(project, breakpoint, editor, pair.first)
+            }
+        }
+        return mutableListOf(action)
+    }
 }